Experimental approach to running `field(converter=...)`, and are-all-fields-present checks, during `__init__`.
This would have been nice as it means that other libraries, like
jaxtyping, could then patch in their own checks into __init__, rather
than the metaclass __call__.

However this approach ends up being a bit of a non-starter, due to the
fact that an end-user may write super().__init__(), which may trigger
checks too early.

Saving as a commit here for posterity, but this should probably not be
merged.